requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.3.0
psycopg-binary==3.1.4
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.2.0